+++ examples/sphere_render.rs
@@ -8,10 +8,10 @@ const CANVAS_HEIGHT: usize = 800;
 
 
 // Original values:
-// const CANVAS_Z: f32 = 10.0;
-// const RAY_ORIGIN_Z: f32  = -5.0;
-const SCREEN_Z: f32 = 60.0;
-const RAY_ORIGIN_Z: f32  = -2.0;
+// const CANVAS_Z: f64 = 10.0;
+// const RAY_ORIGIN_Z: f64  = -5.0;
+const SCREEN_Z: f64 = 60.0;
+const RAY_ORIGIN_Z: f64  = -2.0;
 
 const SPHERE_COLOR: Color = color_rgb!(1.0, 0.2, 1.0);
 
@@ -25,7 +25,7 @@ fn main() {
     let pixel_transform = screen_transform * canvas_transform;
 
     let sphere = Sphere::new(
-        rotation_z(std::f32::consts::FRAC_PI_4) * scaling(1.3, 1.0, 1.0),
+        rotation_z(std::f64::consts::FRAC_PI_4) * scaling(1.3, 1.0, 1.0),
         Material { color: SPHERE_COLOR, ..Material::default() },
     );
     let light = Light::new(point(-10.0, 10.0, -10.0), Color::WHITE);
@@ -34,7 +34,7 @@ fn main() {
     for row in 0..CANVAS_HEIGHT {
         for col in 0..CANVAS_WIDTH {
             xs.clear();
-            let pixel_pos = pixel_transform * point(col as f32, row as f32, 0.0);
+            let pixel_pos = pixel_transform * point(col as f64, row as f64, 0.0);
             let ray = Ray::new(ray_origin,
                                (pixel_pos - ray_origin).normalize());
 
@@ -57,8 +57,8 @@ fn main() {
 }
 
 fn get_canvas_transform() -> M4 {
-    let width_ratio = SCREEN_WIDTH as f32 / CANVAS_WIDTH as f32;
-    let height_ratio = SCREEN_HEIGHT as f32 / CANVAS_HEIGHT as f32;
+    let width_ratio = SCREEN_WIDTH as f64 / CANVAS_WIDTH as f64;
+    let height_ratio = SCREEN_HEIGHT as f64 / CANVAS_HEIGHT as f64;
     scaling(width_ratio, -height_ratio, 0.0)
-        * translation(-(CANVAS_WIDTH as f32) / 2.0, -(CANVAS_HEIGHT as f32) / 2.0, 0.0)
+        * translation(-(CANVAS_WIDTH as f64) / 2.0, -(CANVAS_HEIGHT as f64) / 2.0, 0.0)
 }
